+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_cluster_diff.sql
@@ -1,3 +1,7 @@
 SET search_path = pg_catalog;
 
-ALTER TABLE public.testtable SET WITHOUT CLUSTER;
+DROP INDEX public.mv1_i2;
+
+ALTER TABLE public.t_test_2 SET WITHOUT CLUSTER;
+
+ALTER TABLE public.t_test SET WITHOUT CLUSTER;