+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/modify_cluster_diff.sql
@@ -1,5 +1,7 @@
 SET search_path = pg_catalog;
 
-CREATE INDEX testindex2 ON public.testtable USING btree (f2);
+ALTER TABLE public.t_test CLUSTER ON t_test_i2;
 
-ALTER TABLE public.testtable CLUSTER ON testindex2;
+ALTER TABLE public.t_test_3 CLUSTER ON t_test_3_i;
+
+ALTER MATERIALIZED VIEW public.mv1 CLUSTER ON mv1_i2;